Split f-utils to multiple files
